+++ engine/core/spt-base/src/main/java/com/dell/spt/base/logging/ExtResultsXmlLogMessage.java
@@ -131,7 +131,7 @@ public final class ExtResultsXmlLogMessage extends LogMessageBase {
 						.append(durationSnapshot.max())
 						.append('"');
 
-		strb.append(" />\n");
+		strb.append(" />");
 	}
 
 	private static String escapeXml(final String text) {